refactor: Rename Sorting to Sorter
